Drive the car from Rust; add a calibrate compass command

+++ monitor/status_app.py
@@ -2,6 +2,7 @@
 
 import cherrypy
 import netifaces
+import os.path
 import subprocess
 import sys
 
@@ -111,7 +112,7 @@ class StatusApp(object):
     def run(self):
         """Runs the car."""
         self._check_post()
-        self._command.run_course()
+        self._command.handle_message('start')
         self._logger.info('Received run command from web')
         return {'success': True}
 
@@ -120,7 +121,7 @@ class StatusApp(object):
     def stop(self):
         """Stops the car."""
         self._check_post()
-        self._command.stop()
+        self._command.handle_message('stop')
         self._logger.info('Received stop command from web')
         return {'success': True}
 
@@ -130,17 +131,20 @@ class StatusApp(object):
         """Calibrates the compass."""
         self._check_post()
         self._logger.info('Received calibrate compass command from web')
-        return {'success': False, 'message': 'Not implemented'}
+        self._command.handle_message('calibrate-compass')
+        return {'success': True}
 
     @cherrypy.expose
     @cherrypy.tools.json_out()
     def line_up(self):  # pylint: disable=no-self-use
         """Plays the Mario Kart line up sound."""
         self._check_post()
-        subprocess.Popen(
-            ('mpg123', 'sound/race-start.mp3'),
-            stdout=open('/dev/null', 'w')
-        )
+        self._command.handle_message('line-up')
+        if os.path.isfile('/usr/bin/mpg123') and os.path.isfile('sound/race-start.mp3'):
+            subprocess.Popen(
+                ('/usr/bin/mpg123', 'sound/race-start.mp3'),
+                stdout=open('/dev/null', 'w')
+            )
         return {'success': True}
 
     @cherrypy.expose
@@ -148,10 +152,12 @@ class StatusApp(object):
     def count_down(self):  # pylint: disable=no-self-use
         """Plays the Mario Kart count down sound."""
         self._check_post()
-        subprocess.Popen(
-            ('mpg123', 'sound/count-down.mp3'),
-            stdout=open('/dev/null', 'w')
-        )
+        self._command.handle_message('line-up')
+        if os.path.isfile('/usr/bin/mpg123') and os.path.isfile('sound/count-down.mp3'):
+            subprocess.Popen(
+                ('mpg123', 'sound/count-down.mp3'),
+                stdout=open('/dev/null', 'w')
+            )
         return {'success': True}
 
     @cherrypy.expose
